+++ lib/service_base/active_record_connection_cleaner.rb
@@ -1,0 +1,9 @@
+module ServiceBase
+	class ActiveRecordConnectionCleaner < Middleware
+		def call(env)
+			@app.call(env)
+		ensure
+			ActiveRecord::Base.clear_active_connections!
+		end
+	end
+end