+++ lib/service_base/base_controller.rb
@@ -49,15 +49,15 @@ module ServiceBase
 			def on(routing_key, method_name, message_class=nil)
 				message_class ||= ServiceBase::MessageMap.instance.for(routing_key)
 				raise "Message class not found for the topic: #{routing_key}." unless message_class
-				self.static_message_handlers[routing_key] = MessageHandler.new(message_class, method_name)
+				self.static_message_handlers[routing_key] = MessageRoute.new(message_class, method_name)
 			end
 
 			def unbind(routing_key)
 				self.static_message_handlers.delete(routing_key)
 			end
 
-			def before_setup(name=nil, &block)
-				self.before_setup_block = name ? self.method(name) : block
+			def before_setup(&block)
+				self.before_setup_block = block
 			end
 
 			def setup
@@ -90,12 +90,12 @@ module ServiceBase
 						routing_key = routing_key_prefix + method_name.to_s
 						message_class = ServiceBase::MessageMap.instance.for(routing_key)
 						raise "Topic '#{routing_key}' is not bound to any message. Unable to setup automatic routing for #{self.class}." unless message_class
-						@routing[routing_key] = MessageHandler.new(message_class, method_name)
+						@routing[routing_key] = MessageRoute.new(message_class, method_name)
 					end
 				end
 			end
 
-			def fetch_message_handler(routing_key)
+			def find_message_route(routing_key)
 				raise "Routing not set up" unless @routing
 				@routing[routing_key]
 			end
@@ -127,8 +127,16 @@ module ServiceBase
 			@logger = logger
 		end
 
-		def setup_message(message, delivery_info)
-			@message, @delivery_info = message, delivery_info
+		def process_message(env)
+			@env           = env
+			@action        = env[:action]
+			@message       = env[:message]
+			@delivery_info = env[:delivery_info]
+
+			send @action
+
+		ensure
+			@env, @message, @delivery_info = nil
 		end
 
 		def reply(reply_message, reply_to=nil, message_id=nil)
@@ -146,6 +154,9 @@ module ServiceBase
 																 correlation_id: message_id)
 		end
 
+		class MessageRoute < Struct.new(:message_class, :action)
+		end
+
 		class Dispatcher
 			attr_reader :controller_class, :rabbit_connection, :cache, :logger, :queue
 
@@ -154,6 +165,7 @@ module ServiceBase
 				@rabbit_connection = rabbit_connection
 				@cache             = cache
 				@logger            = logger
+				@application       = ServiceBase.application.to_app
 			end
 
 			def start
@@ -162,53 +174,20 @@ module ServiceBase
 			end
 
 			def process_message(delivery_info, properties, payload)
-				if (message_handler = @controller_class.fetch_message_handler(delivery_info.routing_key))
-					controller_instance = @controller_class.new(@rabbit_connection, @cache, @logger)
-					begin
-						message = message_handler.create_message(delivery_info, properties, payload)
-
-						start_time = Time.now
-						logger.debug "\n\nAMQP Message #{message.class.name} at #{start_time}"
-						logger.debug "Processing by #{@controller_class.name} [Thread: #{Thread.current.object_id}]"
-						logger.debug "\tMessage: #{message.attributes.inspect}."
-						logger.debug "\tProperties: #{properties.inspect}."
-						logger.debug "\tDelivery info: exchange: #{delivery_info[:exchange]}, routing_key: #{delivery_info[:routing_key]}."
-
-						message_handler.call(controller_instance, message, delivery_info)
-
-						logger.debug "Message #{properties[:message_id]} processed in %.1fms." % [(Time.now - start_time) * 1000]
-						true
-					rescue Bunny::Exception
-						# Possible connection error - the controller manager would try to restart connection
-						on_connection_lost $!
-						false
-					end
-				else
-					logger.error "\n\nMessage DROPPED by #{@controller_class.name} at #{Time.now}"
-					logger.error "\tProperties: #{properties.inspect}."
-					logger.error "\tDelivery info: exchange: #{delivery_info[:exchange]}, routing_key: #{delivery_info[:routing_key]}."
-
-					false
-				end
+				env = {
+						delivery_info:     delivery_info,
+						properties:        properties,
+						payload:           payload,
+						controller_class:  controller_class,
+						rabbit_connection: rabbit_connection,
+						cache:             cache,
+						logger:            logger
+				}
+				@application.call(env)
 			rescue Exception => ex
-				logger.error "\n#{ex.class.name} (#{ex.message})"
+				logger.error "\nUNEXPECTED EXCEPTION: #{ex.class.name} (#{ex.message})"
 				logger.error ex.backtrace.join("\n")
-				if properties[:reply_to]
-					begin
-						controller_instance.reply ErrorMessage.new(error: ex.class.name, message: ex.message), properties[:reply_to], properties[:message_id]
-					rescue
-						logger.error "Unable to send error message: #{$!}"
-						logger.error $!.backtrace.join("\n")
-					end
-				end
 				false
-			end
-
-			def on_connection_lost(exception)
-				logger.error "\nConnection lost: #{exception.class.name} (#{exception.message})"
-				logger.error exception.backtrace.join("\n")
-
-				ServiceBase::ControllerManager.stop
 			end
 
 			def create_queue
@@ -225,25 +204,5 @@ module ServiceBase
 				@consumer.cancel if @consumer
 			end
 		end
-
-		class MessageHandler
-			def initialize(message_class, method_name)
-				@message_class = message_class
-				@method_name = method_name
-			end
-
-			def create_message(delivery_info, properties, payload)
-				message = @message_class.new(nil, properties)
-				message.from_json(payload, false)
-				raise InvalidMessageError.new(message.errors.full_messages.join(", "), message) if message.invalid?
-				message
-			end
-
-			def call(controller_instance, message, delivery_info)
-				controller_instance.setup_message(message, delivery_info)
-				controller_instance.send(@method_name)
-			end
-		end
-
 	end
 end