+++ lib/service_base/engine.rb
@@ -8,6 +8,12 @@ module ServiceBase
 			ServiceBase.eager_load_service_messages
 		end
 
+		initializer :use_active_record_connection_release do
+			ActiveSupport.on_load(:active_record) do
+				ServiceBase.application.middlewares.use ServiceBase::ActiveRecordConnectionCleaner
+			end
+		end
+
 		add_engine_migrations
 
 	end